requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
opencv-python>=4.5.0

# Data handling
torchvision>=0.10.0
torch>=1.9.0
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Efficient computing (optional)
faiss-cpu>=1.7.0
numba>=0.56.0

# Progress bars and utilities
tqdm>=4.60.0
joblib>=1.0.0

# Documentation and visualization
jupyter>=1.0.0
ipywidgets>=7.6.0
plotly>=5.0.0

# Development dependencies
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0